:recycle: Feature / Ui 로직 분리
